Progressive Downsampling Initial commit
This feature is there to blur the boundaries of pointclouds, either using an ellipsoid or a polygon. A progressive downsampling is applied based on the distance computed to polygon (projected along z) or ellipsoid. A deterministic uniform random attribute is added to each point and checked against that distance. 
TODO : encapsulate for materials without this downsampling parameters to work, add option to visualiaze density function or to apply it, computations in local coordinate reference frame, non linear density smoothing (sigmoid, log), work with non clockwise polygons, clean code,